perf: adding loaders in buttons and blocking clicks after first click
This avoid that more thant one request is made when some button got clicke

+++ src/pages-components/Stock/components/DeleteItemModal/index.tsx
@@ -1,4 +1,4 @@
-import { SetStateAction, Dispatch, useContext } from 'react';
+import { SetStateAction, Dispatch, useContext, useState } from 'react';
 import { useToast } from '@chakra-ui/react';
 
 import { StockContext } from 'pages-components/Stock';
@@ -12,18 +12,25 @@ type Props = {
 };
 
 export const DeleteItemModal = ({ id, isModalOpen, setIsModalOpen }: Props) => {
+  const [isDeleting, setIsDeleting] = useState(false);
   const { productsDispatch } = useContext(StockContext);
   const toast = useToast();
 
   function handleCloseModal() {
     setIsModalOpen(false);
+    setIsDeleting(false);
   }
 
   async function handleDeleteItem() {
     try {
+      if (isDeleting) {
+        return;
+      }
+      setIsDeleting(true);
       productsDispatch({ type: 'REMOVE-ONE-PRODUCT', payload: { id } });
       const { message } = await StockService.deleteProduct(id as string);
 
+      toast.closeAll();
       toast({
         status: 'success',
         title: message,
@@ -33,9 +40,12 @@ export const DeleteItemModal = ({ id, isModalOpen, setIsModalOpen }: Props) => {
 
       handleCloseModal();
     } catch (error: any) {
+      setIsDeleting(false);
+      toast.closeAll();
       toast({
         status: 'error',
         title: error?.response.data.message,
+        duration: 2000,
       });
     }
   }
@@ -45,6 +55,7 @@ export const DeleteItemModal = ({ id, isModalOpen, setIsModalOpen }: Props) => {
       isModalOpen={isModalOpen}
       handleCloseModal={handleCloseModal}
       handleDeleteItem={handleDeleteItem}
+      isDeleting={isDeleting}
     />
   );
 };
